[proliferate] clarify terminal UI and actions
