+++ scripts/max_lines_allowlist.txt
@@ -47,7 +47,6 @@ desktop/src/components/workspace/files/FileEditorView.tsx
 desktop/src/components/workspace/shell/topbar/HeaderTabs.tsx
 desktop/src/components/workspace/shell/right-panel/RightPanel.tsx
 desktop/src/components/workspace/shell/right-panel/RightPanelHeaderTabs.tsx
-desktop/src/components/workspace/terminals/TerminalPanel.tsx
 desktop/src/hooks/sessions/use-session-control-actions.ts
 desktop/src/hooks/sessions/use-session-creation-actions.ts
 desktop/src/hooks/sessions/use-session-runtime-actions.ts
